修改脚本
