+++ .gitmodules
@@ -3,4 +3,4 @@
 	url = https://github.com/SwiftGGTeam/hexo-theme-casper.git
 [submodule "source"]
 	path = source
-	url = git@github.com:SwiftGGTeam/source.git
+	url = https://github.com/SwiftGGTeam/source.git
