+++ package.json
@@ -8,6 +8,7 @@
   "dependencies": {
     "babel-cli": "^6.24.1",
     "babel-preset-node6": "^11.0.0",
+    "fs-extra": "^5.0.0",
     "hexo": "^3.3.7",
     "hexo-deployer-git": "0.0.4",
     "hexo-generator-archive": "^0.1.2",
@@ -23,6 +24,7 @@
   },
   "devDependencies": {
     "babel-cli": "^6.24.1",
-    "babel-core": "^6.25.0"
+    "babel-core": "^6.25.0",
+    "babel-preset-es2015": "^6.24.1"
   }
 }
